requirements.txt

pybit
python-dotenv